Solve conversation
