+++ deepmd/tf/descriptor/se_atten.py
@@ -1951,7 +1951,7 @@ class DescrptSeAtten(DescrptSeA):
         if self.tebd_input_mode in ["strip"]:
             assert (
                 type(self) is not DescrptSeAtten
-            ), "only DescrptDPA1Compat ansd DescrptSeAttenV2 can serialize when tebd_input_mode=='strip'"
+            ), "only DescrptDPA1Compat and DescrptSeAttenV2 can serialize when tebd_input_mode=='strip'"
             data.update(
                 {
                     "embeddings_strip": self.serialize_network_strip(
